requirements.txt
pandas
requests
pandas-gbq
google-cloud-bigquery
dbt-bigquery